adds transform health status to expanded row in transform list
